enhancement(build) : rename factories
