+++ src/provider/provider.factory.ts
@@ -14,7 +14,7 @@ import {
 import { DeclarationOptions, ModuleDeclarator } from '../utils/module.declarator';
 import { ModuleFinder } from '../utils/module.finder';
 import { Location, NameParser } from '../utils/name.parser';
-import { ProviderOptions } from './schema';
+import { ProviderOptions } from './provider.schema';
 
 export function main(options: ProviderOptions): Rule {
   options = transform(options);
